pass Plugin object to Twig
Instead of creating a new variable to the `templateVar` engine, the plugin
itself is passed to make it easier to design a (Twig) HTML template.

This makes it easy to loop over the pages:

```twig
{% for paginator.pages %}
{% endfor %}
```

instead of the strange looking:

```twig
{% for paginator[paginator.offset] %}
{% endfor %}
```

Also consequently apply this for the previous/next - first/last navigation:

```twig
{% for paginator.previous.pages %}
{% endfor %}

{% for paginator.next.pages %}
{% endfor %}

{% for paginator.first.pages %}
{% endfor %}

{% for paginator.last.pages %}
{% endfor %}
```

Because chaining 'em is awesome!

**Important**

After merge this repo needs to be upgraded to v1.0.0 since it breaks with the
old version.